requirements/requirements_dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements/requirements_dev.txt requirements/requirements.in requirements/requirements_dev.in
#
alabaster==0.7.12
    # via sphinx
attrs==22.1.0
    # via pytest
babel==2.11.0
    # via sphinx
black==22.10.0
    # via -r requirements/requirements_dev.in
bleach==5.0.1
    # via readme-renderer
build==0.9.0
    # via pip-tools
certifi==2022.9.24
    # via requests
cfg-load==0.9.0
    # via -r requirements/requirements.in
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
commonmark==0.9.1
    # via
    #   recommonmark
    #   rich
coverage[toml]==6.5.0
    # via pytest-cov
darglint==1.8.1
    # via -r requirements/requirements_dev.in
docutils==0.17.1
    # via
    #   m2r2
    #   readme-renderer
    #   recommonmark
    #   restructuredtext-lint
    #   sphinx
    #   sphinx-rtd-theme
exceptiongroup==1.0.0
    # via pytest
flake8==5.0.4
    # via
    #   -r requirements/requirements_dev.in
    #   flake8-docstrings
    #   flake8-pyproject
    #   flake8-rst-docstrings
flake8-docstrings==1.6.0
    # via -r requirements/requirements_dev.in
flake8-pyproject==1.1.0.post0
    # via -r requirements/requirements_dev.in
flake8-rst-docstrings==0.2.7
    # via -r requirements/requirements_dev.in
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==5.0.0
    # via twine
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements/requirements_dev.in
jaraco-classes==3.2.3
    # via keyring
jinja2==3.1.2
    # via sphinx
keyring==23.9.3
    # via twine
m2r2==0.3.2
    # via -r requirements/requirements_dev.in
markupsafe==2.1.1
    # via jinja2
mccabe==0.7.0
    # via flake8
mistune==0.8.4
    # via m2r2
more-itertools==9.0.0
    # via jaraco-classes
mpu[io]==0.23.1
    # via cfg-load
mypy==0.982
    # via -r requirements/requirements_dev.in
mypy-extensions==0.4.3
    # via
    #   -r requirements/requirements_dev.in
    #   black
    #   mypy
packaging==21.3
    # via
    #   build
    #   pytest
    #   sphinx
pathspec==0.10.1
    # via black
pep517==0.13.0
    # via build
pip-tools==6.9.0
    # via -r requirements/requirements_dev.in
pkginfo==1.8.3
    # via twine
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
pycodestyle==2.9.1
    # via flake8
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via
    #   flake8-rst-docstrings
    #   readme-renderer
    #   rich
    #   sphinx
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r requirements/requirements_dev.in
    #   pytest-cov
pytest-cov==4.0.0
    # via -r requirements/requirements_dev.in
pytz==2022.6
    # via
    #   babel
    #   cfg-load
    #   mpu
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via cfg-load
readme-renderer==37.3
    # via twine
recommonmark==0.7.1
    # via -r requirements/requirements_dev.in
requests==2.28.1
    # via
    #   cfg-load
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==0.10.1
    # via twine
restructuredtext-lint==1.4.0
    # via flake8-rst-docstrings
rfc3986==2.0.0
    # via twine
rich==12.6.0
    # via twine
six==1.16.0
    # via
    #   bleach
    #   cfg-load
snowballstemmer==2.2.0
    # via
    #   pydocstyle
    #   sphinx
sphinx==5.3.0 ; python_version > "3.7"
    # via
    #   -r requirements/requirements_dev.in
    #   recommonmark
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.1.0
    # via -r requirements/requirements_dev.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
tomli==2.0.1
    # via
    #   -r requirements/requirements.in
    #   black
    #   build
    #   coverage
    #   flake8-pyproject
    #   mypy
    #   pep517
    #   pytest
twine==4.0.1
    # via -r requirements/requirements_dev.in
typing-extensions==4.4.0
    # via mypy
tzdata==2022.6
    # via pytz-deprecation-shim
tzlocal==4.2
    # via mpu
urllib3==1.26.12
    # via
    #   requests
    #   twine
versioneer==0.28
    # via -r requirements/requirements_dev.in
webencodings==0.5.1
    # via bleach
wheel==0.37.1
    # via pip-tools
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
